Format parse utility modules
